+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### [Unreleased]
 
+### Added
+- Allow redeposit expired utxos on bitcoin multisig
+
 ### Changed
 - Improve Assets app cache Behavior
 - Update breez to 0.4.1